- Sim Cache Debugging Support for Stateless Emitters - Adding support so that we can more easily debug what is happening in stateless emitters (what the values are, what attributes are active, etc) - With this support we can also use automation testing to compare the cache results together
#rb rob.krajcarski
#jira UE-219378
#rnx
#tests Capture stateless data inside attribute spreadsheet, etc, and validate that capturing without debugging data during sequencer doesn't pull data in

[CL 35003071 by stu mckenna in ue5-main branch]

+++ Engine/Plugins/FX/Niagara/Source/Niagara/Internal/Stateless/NiagaraStatelessCommon.h
@@ -7,12 +7,12 @@
 #include "NiagaraStatelessCommon.generated.h"
 
 UENUM(meta = (Bitflags, UseEnumValuesAsMaskValuesInEditor = "true"))
-enum class ENiagaraStatelessFeatureMask : uint8
+enum class ENiagaraStatelessFeatureMask : uint32
 {
 	// We can execute on the GPU (Might be broken down into GPUCompute | GPUGraphics | GPUAsyncCompute in future but this will remain the master mask)
-	ExecuteGPU		= 1 << 0,
+	ExecuteGPU			= 1 << 0,
 	// We can execute on the CPU
-	ExecuteCPU		= 1 << 1,
+	ExecuteCPU			= 1 << 1,
 
 	None			= 0							UMETA(Hidden),
 	ExecuteAll		= ExecuteGPU | ExecuteCPU	UMETA(Hidden),
